moved conversation state to hook
